fixed some small bugs and got rid of some warnings
Ignore-this: 8e32f6d5fb2b19b4b522fb0e1bc0d754

darcs-hash:20131112051551-c21fe-cb9ab0fbb9614a79fcf4eb7c8bb60d13011ba4d5

+++ src/network/stungun.cpp
@@ -370,7 +370,7 @@ namespace fire
             {
                 uint8_t* ip_p = reinterpret_cast<uint8_t*>(&ip6[IPV6_LENGTH-1]);
                 std::stringstream ips;
-                for(size_t i = IPV6_LENGTH-2; i >= 0; i--) ips << ip_p[i];
+                for(int i = IPV6_LENGTH-2; i >= 0; i--) ips << ip_p[i];
                 mapped.ip = ips.str();
             }
